+++ CHANGELOG.md
@@ -13,6 +13,12 @@ will not have contiguous patch numbers. Initial major and minor releases will be
 in this file without a patch number. Patch version will be included for bug fixes, but
 may not exactly match a publicly released version.
 
+## [Unreleased]
+
+### Changed
+* Always generate debug info, regardless of configuration property, unless `NeoCscDebugInfo` property is false (#37)
+
+
 ## [3.3.23] - 2022-07-06
 
 ### Added